[HttpClient] Add `DecoratorTrait` to ease writing simple decorators
