+++ src/Symfony/Component/HttpClient/AsyncDecoratorTrait.php
@@ -13,7 +13,6 @@ namespace Symfony\Component\HttpClient;
 
 use Symfony\Component\HttpClient\Response\AsyncResponse;
 use Symfony\Component\HttpClient\Response\ResponseStream;
-use Symfony\Contracts\HttpClient\HttpClientInterface;
 use Symfony\Contracts\HttpClient\ResponseInterface;
 use Symfony\Contracts\HttpClient\ResponseStreamInterface;
 
@@ -24,12 +23,7 @@ use Symfony\Contracts\HttpClient\ResponseStreamInterface;
  */
 trait AsyncDecoratorTrait
 {
-    private $client;
-
-    public function __construct(HttpClientInterface $client = null)
-    {
-        $this->client = $client ?? HttpClient::create();
-    }
+    use DecoratorTrait;
 
     /**
      * {@inheritdoc}
@@ -50,16 +44,5 @@ trait AsyncDecoratorTrait
         }
 
         return new ResponseStream(AsyncResponse::stream($responses, $timeout, static::class));
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function withOptions(array $options): self
-    {
-        $clone = clone $this;
-        $clone->client = $this->client->withOptions($options);
-
-        return $clone;
     }
 }
